write record with execution name and rename to reflect new scope
