+++ packages/api/webpack.config.js
@@ -48,7 +48,7 @@ module.exports = {
     sfSemaphoreDown: './lambdas/sf-semaphore-down.js',
     sfStarter: './lambdas/sf-starter.js',
     sqsMessageConsumer: './lambdas/sqs-message-consumer.js',
-    writeSQStoS3: './lambdas/write-sqs-to-s3.js',
+    writeDbDlqRecordstoS3: './lambdas/write-db-dlq-records-to-s3.js',
   },
   devtool: 'inline-source-map',
   resolve: {
